Rename minimal package to server
