+++ command/monitor/monitor.go
@@ -9,7 +9,7 @@ import (
 	"syscall"
 
 	"github.com/0xPolygon/polygon-sdk/command/helper"
-	"github.com/0xPolygon/polygon-sdk/minimal/proto"
+	"github.com/0xPolygon/polygon-sdk/server/proto"
 	empty "google.golang.org/protobuf/types/known/emptypb"
 )
