+++ command/peers/peers_add.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/0xPolygon/polygon-sdk/command/helper"
 	helperFlags "github.com/0xPolygon/polygon-sdk/helper/flags"
-	"github.com/0xPolygon/polygon-sdk/minimal/proto"
+	"github.com/0xPolygon/polygon-sdk/server/proto"
 )
 
 // PeersAdd is the PeersAdd to start the sever
