+++ e2e/discovery_test.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/0xPolygon/polygon-sdk/e2e/framework"
-	"github.com/0xPolygon/polygon-sdk/minimal/proto"
+	"github.com/0xPolygon/polygon-sdk/server/proto"
 	empty "google.golang.org/protobuf/types/known/emptypb"
 )
