+++ server/system_service.go
@@ -1,11 +1,11 @@
-package minimal
+package server
 
 import (
 	"context"
 	"time"
 
-	"github.com/0xPolygon/polygon-sdk/minimal/proto"
 	"github.com/0xPolygon/polygon-sdk/network"
+	"github.com/0xPolygon/polygon-sdk/server/proto"
 	"github.com/libp2p/go-libp2p-core/peer"
 	empty "google.golang.org/protobuf/types/known/emptypb"
 )
